% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookstuff_function.R
\name{bookStuff}
\alias{bookStuff}
\title{bookstuff}
\usage{
bookStuff()
}
\arguments{
\item{None}{This function does not take any parameters}
}
\value{
a shiny app
}
\description{
function creates a shiny app which lets you manage your book list and export it as a csv.
}
